log add

+++ src/main/java/FunFictionUserProject/funFictionUser/controller/AdminController.java
@@ -2,11 +2,9 @@ package FunFictionUserProject.funFictionUser.controller;
 
 import FunFictionUserProject.funFictionUser.dto.UserListDto;
 import FunFictionUserProject.funFictionUser.exeption.EntityNotFoundException;
-import FunFictionUserProject.funFictionUser.service.FunFictionService;
-import FunFictionUserProject.funFictionUser.service.GenreService;
-import FunFictionUserProject.funFictionUser.service.RoleService;
-import FunFictionUserProject.funFictionUser.service.UserService;
+import FunFictionUserProject.funFictionUser.service.*;
 import FunFictionUserProject.funFictionUser.view.*;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,6 +18,7 @@ import java.util.List;
 import static FunFictionUserProject.funFictionUser.util.DataConstant.*;
 import static FunFictionUserProject.funFictionUser.util.UrlConstant.*;
 
+@Slf4j
 @RestController
 @RequestMapping(ADMIN)
 public class AdminController {
@@ -29,13 +28,15 @@ public class AdminController {
     private List<Role> userRoles = new ArrayList<>();
     private final FunFictionService funFictionService;
     private final GenreService genreService;
+    private final CommentsService commentsService;
 
     @Autowired
-    public AdminController(UserService userService, RoleService roleService, FunFictionService funFictionService, GenreService genreService) {
+    public AdminController(UserService userService, RoleService roleService, FunFictionService funFictionService, GenreService genreService, CommentsService commentsService) {
         this.userService = userService;
         this.roleService = roleService;
         this.funFictionService = funFictionService;
         this.genreService = genreService;
+        this.commentsService = commentsService;
     }
 
     /*
@@ -44,13 +45,15 @@ public class AdminController {
     @PostMapping(SAVE)
     public ResponseEntity<User> save(@RequestBody User user, HttpSession session) {
         User userCheck = userService.findByLogin(user.getLogin());
+        log.info("user find by login method save in adminController");
         if (userCheck != null) {
-            throw new EntityNotFoundException(User.class, user);
+            throw new EntityNotFoundException("user not found");
         }
         user.setStatus(Status.ACTIVE);
         user.setCreated(new Date());
         user.setUpdated(new Date());
         User userResult = userService.registerUser(user);
+        log.info("user registerUser by login method save in adminController");
         session.setAttribute(USER_RESULT, userResult);
         return new ResponseEntity<>(userResult, HttpStatus.CREATED);
     }
@@ -58,26 +61,33 @@ public class AdminController {
     /*
     функция для назначения админом со списка юзеров (8)
      */
-    @GetMapping(UPDATE_ADMIN)
-    public ResponseEntity<User> updatedUserByAdmin(@PathVariable(ID) Long id) {
+    @GetMapping("/updatedUserByAdmin/{id}")
+    public ResponseEntity<UserListDto> updatedUserByAdmin(@PathVariable(ID) Long id) {
         User userSearch = userService.findById(id);
+        log.info("user findById method updatedUserByAdmin in adminController");
+        if (userSearch == null) {
+            throw new EntityNotFoundException("user not found");
+        }
         Role role = roleService.findByRole(ROLE_ADMIN);
+        log.info("role findByRole method updatedUserByAdmin in adminController");
         userRoles.add(role);
         userSearch.setRoles(userRoles);
-        userService.saveAndFlush(userSearch);
-        return new ResponseEntity<>(userSearch, HttpStatus.OK);
+        User updatedUser = userService.saveAndFlush(userSearch);
+        log.info("user saveAndFlush method updatedUserByAdmin in adminController");
+        UserListDto userListDto = UserListDto.fromUser(updatedUser);
+        return new ResponseEntity<>(userListDto, HttpStatus.OK);
     }
 
     /*
     лист всех пользователей со страницы админа (5)
      */
-    @GetMapping(LIST_USER)
+    @GetMapping("/allUserByAdmin")
     public ResponseEntity<List<UserListDto>> getAllByUser() {
         List<User> users = userService.findAll();
+        log.info("user list findAll method getAllByUser in adminController");
         List<UserListDto> userList = new ArrayList<>();
         for (User user : users) {
-            UserListDto userListDto = new UserListDto();
-            userList.add(userListDto.fromUser(user));
+            userList.add(UserListDto.fromUser(user));
         }
         return new ResponseEntity<>(userList, HttpStatus.OK);
     }
@@ -85,22 +95,43 @@ public class AdminController {
     /*
     удаление юзера со страницы админа (6)
      */
-    @DeleteMapping(DELETE_USER)
-    public ResponseEntity<User> deleteUserByAdmin(@PathVariable(ID) Long id) {
+    @DeleteMapping("/deleteUserByAdmin/{id}")
+    public ResponseEntity<List<UserListDto>> deleteUserByAdmin(@PathVariable(ID) Long id) {
         User user = userService.findById(id);
-        userService.delete(user);
-        return new ResponseEntity<>(HttpStatus.OK);
+        log.info("user findById method deleteUserByAdmin in adminController");
+        if (user == null) {
+            throw new EntityNotFoundException("user not found");
+        }
+        user.setStatus(Status.DELETE);
+        userService.saveAndFlush(user);
+        log.info("user saveAndFlush method deleteUserByAdmin in adminController");
+        List<User> userList = userService.findAll();
+        List<UserListDto> userDtoList = new ArrayList<>();
+        for (User users : userList) {
+            userDtoList.add(UserListDto.fromUser(users));
+        }
+        return new ResponseEntity<>(userDtoList, HttpStatus.OK);
     }
 
     /*
     блокировка юзера со страницы админа (7)
     */
-    @DeleteMapping(BLOCK_USER)
-    public ResponseEntity<User> blockUserByAdmin(@PathVariable(ID) Long id) {
+    @GetMapping("/blockUserByAdmin/{id}")
+    public ResponseEntity<List<UserListDto>> blockUserByAdmin(@PathVariable(ID) Long id) {
         User user = userService.findById(id);
+        log.info("user findById method blockUserByAdmin in adminController");
+        if (user == null) {
+            throw new EntityNotFoundException("user not found");
+        }
         user.setStatus(Status.BLOCK);
-        User userResult = userService.saveAndFlush(user);
-        return new ResponseEntity<>(userResult, HttpStatus.OK);
+        userService.saveAndFlush(user);
+        log.info("user saveAndFlush method blockUserByAdmin in adminController");
+        List<User> userList = userService.findAll();
+        List<UserListDto> userDtoList = new ArrayList<>();
+        for (User users : userList) {
+            userDtoList.add(UserListDto.fromUser(users));
+        }
+        return new ResponseEntity<>(userDtoList, HttpStatus.OK);
     }
 
     /*
@@ -109,6 +140,10 @@ public class AdminController {
     @GetMapping(INFO_USER)
     public ResponseEntity<User> infoUserByAdmin(@PathVariable(ID) Long id) {
         User user = userService.findById(id);
+        log.info("user findById method infoUserByAdmin in adminController");
+        if (user == null) {
+            throw new EntityNotFoundException("user not found");
+        }
         return new ResponseEntity<>(user, HttpStatus.OK);
     }
 
@@ -118,7 +153,12 @@ public class AdminController {
     @DeleteMapping(DELETE_FUN_FIC_ADMIN)
     public ResponseEntity<FunFiction> deleteFunFunficByAdmin(@PathVariable(ID) Long id) {
         FunFiction funFiction = funFictionService.findById(id);
+        log.info("funFiction findById method deleteFunFunficByAdmin in adminController");
+        if (funFiction == null) {
+            throw new EntityNotFoundException("fun fic not found");
+        }
         funFictionService.delete(funFiction);
+        log.info("funFiction delete method deleteFunFunficByAdmin in adminController");
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
@@ -128,11 +168,17 @@ public class AdminController {
     @PostMapping(UPDATE_FUN_FIC_ADMIN)
     public ResponseEntity<FunFiction> updatedFunFunficByAdmin(@RequestBody FunFiction funFiction) {
         FunFiction funFictionResult = funFictionService.findById(funFiction.getId());
+        log.info("funFiction findById method updatedFunFunficByAdmin in adminController");
+        if (funFictionResult == null) {
+            throw new EntityNotFoundException("fun fic not found");
+        }
         funFictionResult.setNameFun(funFiction.getNameFun());
         funFictionResult.setShortDescription(funFiction.getShortDescription());
         Genre genre = genreService.findByTypeGenre(funFiction.getGenre().getTypeGenre());
+        log.info("genre findByTypeGenre method updatedFunFunficByAdmin in adminController");
         funFictionResult.setGenre(genre);
         funFictionService.saveAndFlush(funFictionResult);
+        log.info("genre saveAndFlush method updatedFunFunficByAdmin in adminController");
         return new ResponseEntity<>(funFictionResult, HttpStatus.OK);
     }
 }
